agentapi-last/agent/requirements.txt
fastapi
uvicorn
python-dotenv
openai
pymupdf
python-docx
pandas
openpyxl
opencv-python
moviepy
python-multipart
